refactor(layout): split compositor/ and palette/ to top-level peers of frontend/
Pulls `compositor/` and `palette/` out of `frontend/` and parks them
as crate-level peers next to `input/`, `lua/`, `map/`, etc. Both are
heavyweight subsystems with their own internal structure that
multiple peers reach into (lua/bridge implements `Component` and
`PaletteProvider`; input/keymap dispatches keys against the
compositor stack); top-level placement matches the actual dependency
graph and frees `frontend/` to be just the App / runner / dispatch.

Mechanics:

- `git mv src/frontend/compositor src/compositor`
- `git mv src/frontend/palette    src/palette`
- `lib.rs`: add `pub mod compositor;` and `pub mod palette;`
- `frontend/mod.rs`: drop the `pub(crate) mod compositor;` and
  `pub mod palette;` declarations
- Mass `crate::frontend::compositor` → `crate::compositor` and
  `crate::frontend::palette` → `crate::palette` rename across `src/`
- One stray `ttymap::frontend::palette::install` in `main.rs` rewired
- `Compositor::is_empty` added so `len()` no longer trips clippy's
  `len_without_is_empty` lint at the new public surface
- Doc paths in `CLAUDE.md`, `README.md`, `docs/*.md` updated to
  reflect the new locations

No behaviour change. 388 / 388 tests pass; clippy / fmt clean.

After this PR:

```
src/
├── frontend/        runner + dispatch + UserIntent + AppEvent + ui shell
├── compositor/      stack-based focus/modal system (top-level peer)
├── palette/         `:`-triggered universal picker (top-level peer)
├── input/           keymap + mouse adapter + input thread
├── lua/             plugin runtime + bridge
├── map/             domain
├── theme/, geo/, shared/, commands/, config.rs, logging.rs
```

Lines moved, not added — the diff is dominated by the directory
moves and the import path rewrites; no logic changed.
