+++ src/compositor/sidebar.rs
