+++ src/palette/action.rs
@@ -9,8 +9,8 @@
 //!
 //! [`PaletteProvider`]: super::provider::PaletteProvider
 
+use crate::compositor::Component;
 use crate::frontend::UserIntent;
-use crate::frontend::compositor::Component;
 
 use super::provider::PaletteProvider;
